Add classical value bit slicing on int / uint
Implements the OpenQASM 3 classical value bit slicing rules for `int[n]`
and `uint[n]`, as reads and as assignment targets, including elements of
arrays of integers.

Index 0 of an integer is its least-significant bit, the opposite of the
`bit[n]` register convention, where bit 0 is the most-significant bit of
the stored value. The two conventions are converted at the boundary:
`read_int_bits` / `write_int_bits` work in LSB-first integer space and
hand back a width-carrying `BitValue`.

Three sites changed:

- `Qasm3Analyzer.resolve_index` is extracted from
  `analyze_classical_indices`, so a single subscript can be resolved
  against a bit width rather than an array dimension. The extraction is
  behaviour-preserving for the container path; `normalize_index` gains
  an optional `type_name` so an out-of-range bit index names the
  declared width: `Index 32 out of range for 'int[32]' variable 'myInt'`.
- `Qasm3ExprEvaluator._get_var_value` takes an integer bit-view path
  before the container path.
- `_visit_classical_assignment` detects a bit-slice lvalue up front and
  bypasses the scalar cast, which is what produced
  `Cannot cast 'str' to 'IntType'`.

A chained subscript is split as element indices plus one trailing bit
subscript rather than flattened, so `intArr[0][0] = 1` and
`a[0][0][3] = 1` resolve instead of failing with
`Invalid index for variable`.

A descending range needs an explicit negative step. `myInt[-1:-16]` is
an empty index set under the spec's normative range definition, even
though the spec's own example assigns it to a `bit[16]`; the ambiguous
form is rejected and `myInt[-1:-1:-16]` gives the reversed slice.

The `indexing_non_array` fixture moves from `int` to `float`, because
`int x; x[0] = 4;` is now a valid lvalue with an out-of-range value.

+++ tests/qasm3/resources/variables.py
@@ -297,12 +297,14 @@ ASSIGNMENT_TESTS = {
         8,
         "float[32] x = 1.2345678912345679e+44;",
     ),
+    # ``int`` / ``uint`` scalars are excluded here: they are bit-sliceable per the
+    # OpenQASM classical value bit slicing rules, so ``x[0]`` is a valid lvalue.
     "indexing_non_array": (
         """
         OPENQASM 3.0;
         include "stdgates.inc";
 
-        int x = 3;
+        float x = 3.0;
         x[0] = 4;
         """,
         "Invalid index for variable 'x'",
